Code for ackermann control
Contains the code to control the movement of the rover from inputs of 1/curve radius and rover speed and outputs motor rotation speed and rotation angle for each of the 6 wheels. Speed and angle are restricted based on certain conditions such as when the radius is inside the rover's body and the maneuver will calculate it as if the radius was at the edge of the rover. Speed is limited if one motor exceeds maximum speed so can complete ackermann maneuver properly.